requirements.txt
NodeGraphQt
PySide6
pycryptodome
